+++ client/src/utils/date/getDayOfWeek.ts
