+++ client/src/utils/date/getTripPeriod.ts
