+++ client/src/utils/date/getTripTitleWithRegion.ts
