+++ client/src/utils/date/index.ts
@@ -1,0 +1,4 @@
+export { default as getDayOfWeek } from './getDayOfWeek';
+export { default as getFormatDateString } from './getFormatDateString';
+export { default as getTripPeriod } from './getTripPeriod';
+export { default as getTripTitleWithRegion } from './getTripTitleWithRegion';
